requirements.txt
flask
gradio
google-generativeai
python-dotenv
flask==3.0.0
google-generativeai==0.7.0
python-dotenv==1.0.0